% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WNBAPerGameAdvStatistics.R
\name{WNBAPerGameAdvStatistics}
\alias{WNBAPerGameAdvStatistics}
\title{WNBA Player Advanced Statistics For a Given Season}
\usage{
WNBAPerGameAdvStatistics(season = 2018)
}
\arguments{
\item{season}{A numeric year}
}
\value{
An object of class tbl_df containing all player advanced statistics for each player from a WNBA season on basketball-reference.com.
}
\description{
WNBA Player Advanced Statistics For a Given Season
}
\examples{
WNBAPerGameAdvStatistics(2018)

}
