% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WNBAPerGameStatistics.R
\name{WNBAPerGameStatistics}
\alias{WNBAPerGameStatistics}
\title{WNBA Player Statistics For a Given Season}
\usage{
WNBAPerGameStatistics(season = 2018)
}
\arguments{
\item{season}{A numeric year}
}
\value{
An object of class tbl_df containing all player statistics for each player from a WNBA season on basketball-reference.com.
}
\description{
WNBA Player Statistics For a Given Season
}
\examples{
WNBAPerGameStatistics(2018)
}
