% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WNBAPerGameStatisticsPer36Min.R
\name{WNBAPerGameStatisticsPer36Min}
\alias{WNBAPerGameStatisticsPer36Min}
\title{WNBA Player Statistics For a Given Season per 36 Minutes}
\usage{
WNBAPerGameStatisticsPer36Min(season = 2018)
}
\arguments{
\item{season}{A numeric year}
}
\value{
An object of class tbl_df containing all player statistics per 36 minutes for each player from a WNBA season on basketball-reference.com.
}
\description{
WNBA Player Statistics For a Given Season per 36 Minutes
}
\examples{
WNBAPerGameStatisticsPer36Min(2018)
}
