% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WNBAPlayerPerGameStats.R
\name{WNBAPlayerPerGameStats}
\alias{WNBAPlayerPerGameStats}
\title{WNBA Player Career Statistics}
\usage{
WNBAPlayerPerGameStats(player_link)
}
\arguments{
\item{player_link}{The end of url of 'https://www.basketball-reference.com/wnba'. For example, '/players/f/fowlesy01w.html' for Sylvia Fowles.}
}
\value{
An object of class tbl_df containing a player's career statistics from basketball-reference.com.
}
\description{
WNBA Player Career Statistics
}
\examples{
WNBAPlayerPerGameStats('/players/f/fowlesy01w.html')
}
