% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WNBASeasonTeamByYear.R
\name{WNBASeasonTeamByYear}
\alias{WNBASeasonTeamByYear}
\title{Season and Team Stats for WNBA team}
\usage{
WNBASeasonTeamByYear(team, season = 2018)
}
\arguments{
\item{team}{A team of form "MIN", "DET", etc. Go to 'https://www.basketball-reference.com/wnba/teams/' to find extension needed.}

\item{season}{A numeric year}
}
\value{
An object of class tbl_df containing team summary statistics for a season. Note this is a different result than 'NBASeasonTeamByYear' in ballr package because basketball-reference.com provides different information between NBA and WNBA.
}
\description{
Season and Team Stats for WNBA team
}
\examples{
WNBASeasonTeamByYear(team = "Min", season = 2018)
}
