% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WNBAStandingsBySeason.R
\name{WNBAStandingsBySeason}
\alias{WNBAStandingsBySeason}
\title{Standings of WNBA season standings}
\usage{
WNBAStandingsBySeason(season = 2018)
}
\arguments{
\item{season}{A numeric year}
}
\value{
A list that contains the Eastern and Western Conference WNBA standings for a season.
}
\description{
Standings of WNBA season standings
}
\examples{
WNBAStandingsBySeason(2018)
}
